requirements.txt
numpy
pillow
pyqt6
maturin
